+++ app/models/inventory.rb
@@ -7,8 +7,8 @@ class Inventory < ApplicationRecord
   include Storage
 
   def prompt_to_employee(employee, quantity)
-    update!(quantity: self.quantity.to_i - quantity.to_i)
-    employee.inventory_prompts.create!(product: self.product, quantity: quantity.to_i)
+    update!(quantity: self.quantity.to_f - quantity.to_f)
+    employee.inventory_prompts.create!(product: self.product, quantity: quantity.to_f)
   end
 
   def request_for_inventory(requestor_employee, quantity_asked, date_of_use)
