+++ app/models/inventory_prompt.rb
@@ -5,7 +5,7 @@ class InventoryPrompt < ApplicationRecord
   def accept!
     update!(is_accepted: true)
     assigned_inventory = employee.employees_inventories.find_or_create_by(product: product)
-    assigned_inventory.update!(quantity: (assigned_inventory.quantity.to_i + self.quantity.to_i))
+    assigned_inventory.update!(quantity: (assigned_inventory.quantity.to_f + self.quantity.to_f))
   end
 
   def reject!
@@ -16,7 +16,7 @@ class InventoryPrompt < ApplicationRecord
                   Employee.find_by(name: assigned_by).employees_inventories
                 end.find_or_create_by(product: product)
 
-    assignor.update!(quantity: assignor.quantity.to_i + quantity.to_i)
+    assignor.update!(quantity: assignor.quantity.to_f + quantity.to_f)
 
     destroy!
   end
